Add support for dropdown entry
It is easier and more ergonomic if the user can use the keyboard in a 
sensible way when working with inputs for large amounts of data entry.

Here the change provides the defaults to the various lists, and a check 
on the validity of a given addition (because locations cannot be 
'Complete' multiple times).