Add a metaprogrammed stream constructor
Given that many event emitters emit multiple types of events, it helps 
to not have to `fromEvents` every single event of interest.  This change 
creates a helper that uses a trailing `$` convention on a property to 
indicate that you want an event stream constructed around the name of 
that stream as a property.

The code has been changed to reflect this utility.